Move icons into common folder
