+++ frontend/src/features/common/icons/CoordinatorIcon.tsx
