+++ frontend/src/features/common/icons/HostIcon.tsx
