+++ frontend/src/features/common/icons/InProgressIcon.tsx
