+++ frontend/src/features/common/icons/index.ts
